Add `pydantic_v1` module to make Olive `pydantic` version agnostic (#788)
## Describe your changes
Pydantic v2 had significant breaking changes which blocks migration to
it. So, we pinned the version to v1.

Pydantic v2 has a `v1` namespace inside it. 

This PR introduces `olive.common.pydantic_v1` module that handles the
import of `v1` api so that Olive can be version agnostic wrt pydantic.
This way, we won't have version conflict with client applications such
as that in https://github.com/microsoft/Olive/issues/709.

Note:
The doc builder still needs pydantic v1 and autodoc_pydantic v1 since we
are using the v1 api.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link