Fix onnx conversion of bnb quantized model, Remove static methods from `HfConfigMixin`  (#807)
## Describe your changes
This PR fixes some issues with the conversion of 4bit quantized models
to onnx:
- The adapters are unloaded first so that the correct quantized moduels
can be found
- A new pytorch handler is created. Previously, we were ignoring the
adapters

Changes the static method in `HfConfigMixin` to regular methods since
these were made static to support the above use case. But they are no
longer required to be static. The code is now more consistent.

Test requirements updated to for transformers since lora unit tests fail
otherwise.
 
## Checklist before requesting a review
- [x] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [x] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link