Add `LoftQ` pass (#861)
## Describe your changes
This PR adds a new pass `LoftQ` which is based on the paper
https://arxiv.org/abs/2310.08659.
It adds on the qlora fine tuning approach by initializing the
quantization and lora adapters together. This "alleviates the
discrepancy between the quantized and full-precision model and
significantly improves generalization in downstream tasks".

Since most of the fine-tuning steps are the same as `QLoRA`, the common
code has been abstracted into a new base class called `QLoRABase`.
`LoftQ` is a separate pass and not an option for `QLoRA` so that we
don't conflate it too much and make the configuration confusing for the
user. They can just switch the pass type instead.

 
## Checklist before requesting a review
- [x] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [x] Update documents if necessary.
- [x] Lint and apply fixes to your code by running `lintrunner -a`
- [x] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

Introduced a new pass `LoftQ` which performs model fine-tuning using the
LoftQ initialization proposed in https://arxiv.org/abs/2310.08659.

## (Optional) Issue link