🪟 snpe windows case (#945)
## Describe your changes
To complete the SNPE support in windows. 
Previously Olive only support to run SNPE toolkit in windows. In this
PR, I tried to:
1. Test the SNPE toolkit in windows and add the required env variables
for SNPE to run. Several differences between run in windows and linux:
- In windows, only the executable file can be find if it's parent folder
in system PATH. That need olive to resolve the full file path to run for
those snpe-tools which are python file but not .exe.
- In windows, to ensure Python.exe can be initialized successfully,
olive need to copy the SYSTEMROOT and other env variables to run the
cmd.
https://stackoverflow.com/questions/58997105/fatal-python-error-failed-to-get-random-numbers-to-initialize-python
2. Enable windows snpe toolkit e2e tests case with Inception model.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.

## (Optional) Issue link