Resumption from checkpoint in LoRA based passes (#988)
## Describe your changes
This PR adds support for resuming training in LoRA based passes. Like in
transformers and ort-training examples, the training can be resumed by
one of:
- Provide a valid checkpoint directing using `resume_from_checkpoint`
option
- Provide an `output_dir`. Checkpoints will be saved in this directory
instead of a temporary directory. If the pass is run again with the same
`output_dir`, it will find the last checkpoint in the directory and
resume training from it.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [x] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
`LoRA`, `QLoRA` and `LoftQ` passes support resuming training from
checkpoints.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link