Check model size in `model_proto_to_file` (#1069)
## Describe your changes
If `save_as_external_data` is `False` we try to save the model and then
with `True` if it fails. However, this is inefficient. It also leads to
a corrupted model file in python 3.11.

In this PR:
- We instead check the model size and set `save_as_external_data` to
`True` if it is negative or above the max protobuf size.
- Check for max size is same as that in onnx that causes the error that
we were previously trying to catch
https://github.com/onnx/onnx/blob/4128a09009aa67622c6308c82fe4199813a71682/onnx/serialization.py#L102
- Negative values can happen on Windows and it is a known issue
https://github.com/onnx/onnx/issues/5861. Code on the onnxruntime
repository also check for this in
https://github.com/microsoft/onnxruntime/blob/0acde1157afc9f6cf278a8f97cc3da6f048a649a/onnxruntime/python/tools/transformers/models/stable_diffusion/ort_optimizer.py#L114
- I originally added the try/except logic because of the negative model
size issue #257
- Changed the log from warning to info since it doesn't require any
change from the user. The warning also causes confusion making users
think there was a problem in the pass.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link
#1023
#926