Restrict SliceGPT pass to only supported datasets
SliceGPT supports only a limited number of datasets for
calibration purposes. However, the Olive/SliceGPT pass' use of
DataConfig to accept that user parameter gave the impression that
any/all datasets are legal.

Removed use of DataConfig and instead accept a dataset name (from
among the list of supported datasets) to resolve the ambiguity.