Make the accelerator EP optional for non-onnx pass (#1072)
## Describe your changes
* Originally, the execution_providers in accelerator spec is required
regardless whether it is used or not. Even if user doesn't specify them,
the local system will infer and get the available EPs to autofill the
EPs. This is painful for passes that doesn't use the EP.
* In this PR, if there is no pass that belongs to onnx techniques, we
will not infer the EP based on the device.
* On the other hand, if the device is not specified, we will choose cpu
by default.
## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link