Bug fix: Don't fail `get_hf_model_io_config`, Optional Accelerator execution_provider (#1116)
## Describe your changes
This fixes some bugs introduced by some recent PRs:
- #1091
- made `model.io_config` a property which calls `get_hf_model_io_config`
if the conditions meet.
- However, this function fails if the model is not supported. Now it
doesn't fail and logs when unsupported.
- `PytorchModelHandler.to_json` uses `_io_config` for `"io_config"` so
that the serialized model is the same as the one that was initialized.
- #1072
- execution_provider is optional for non-onnx pass but `to_json` still
has the key. Which doesn't match the signature for
https://github.com/microsoft/Olive/blob/91c4c20e324fbc47dcbbea20bf89035ec94fa4a6/olive/passes/olive_pass.py#L459
   - Only add when not None
- Also removed `pass_accelerator_type` and `pass_execution_provider`
from the aml_pass_runner command line inputs. These values are already
present in `pass_config.json`
 
## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link