Fix handling of extra_args for ModelBuilder pass
GenAI model builder's extra_args works based on existence of a key
rather then the value of the key itself. So, Olive cannot pass a key
with a value of true/false. The key has to explicitly be added/removed
based on the value of the key itself in the user config.