AzureMLSystem: Unique input resources (#1237)
## Describe your changes
- When the same AzureML model resource is provided as input to a command
multiple times, it leads to an invalid job (interestingly, this is not
the case for a local resource).
- AzureML system is updated so that a particular resource is provided as
input only once. The resources are now referenced using names of format
`resource__{i}`. AML runners use the `num_resources` and resource maps
to reconstruct the configs.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link