CI: Only run component governance in package publish pipeline (#1250)
## Describe your changes
- Remove component governance step from the test pipelines. We don't
publish anything in these pipelines, so it is not necessary. This task
is flaky sometimes and fails due to out of disk error. There is no
option to set a different directory to fix this.
- Added the step to the `package_publish` pipeline which is used to
generate the wheel for pypi release. Ignore the docs and examples
directories since those are not part of the package.
- Added the pipeline specific variables in the pipeline yamls instead of
the job templates for easier maintenance.
 
## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link