LoRA: Remove `eval_dataset_size`, Generalize data collator, Remove redundant examples (#1308)
## Describe your changes
- Remove `eval_dataset_size` option which was used to split train data
into train and eval. this is a specific use case which is supported
using data configs that use slices of the train data. We want to remove
as much of the data processing logic from the pass to make it generic
- Data collator is kept for now since the trainer has its own batch size
(which depends on batch size params as well as number of gpus). It is
more generic now.
- olive `Dataset.to_hf_dataset` removed. The logic has been improved and
moved to the pass since it's specific to it.
- Remove `phi` and `open_llama` qlora examples since these are copies of
the examples for `llama`, `phi2` and `phi3`. Keep multiple copies of
essentially the same workflow is hard to maintain when we make changes.
- Other lora examples updated to remove redundant options.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link