Conversion: Absorb lora scaling into lora_B weights (#1329)
## Describe your changes
LoRA has a scaling factor that is computed based on `lora_alpha` and
`r`. These get multiplied with the output of the lora matmuls which
results in an additional `Mul` node with the scaling float. However,
this mul node is hard to keep track of during subsequent optimization
steps (for example, it might get removed if scaling is 1). We also want
to be able to use different sets of adapters which might not share the
same scaling factor.

To overcome this, we implement a context manager which finds all lora
linear layers and wraps it with a custom module class that already has
the scaling multiplied into the lora_B weights. This way, we can just
ship the scaled weights.

Also moved other peft related model transformations into the context
manager.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link