ExtractAdapters: `dynamic_lora_r` and `optional_inputs` (#1326)
## Describe your changes
ONNXModelHandler:
- New `change_model_path_to_dir` method. Currently, there are two
different places where this logic is implemented manually. In onnx
conversion, it doesn't update the model_file_name which leads to warning
when resolving the onnx model file path

ExtractAdapter:
-  `dynamic_lora_r`: Option to make the dimension for LoRA r static or
dynamic.
- `optional_inputs`: Onnx and ort allow the same name to be used for
graph initializers and inputs. When this happens, the input behaves like
an optional input where the initializer is used if the input is not
provided. https://github.com/onnx/onnx/blob/main/docs/IR.md#nodes
Using this behavior, we can set the default initializer for the lora
weights to be empty tensors with shapes `[in_dim, 0]` and `[0,
out_dim]`. When the lora weights are not provided during session run,
these initializers would be used which is equivalent to running the base
model with no adapters.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link