ModelBuilder: Use `TRANSFORMERS_CACHE` as `cache_dir` (#1345)
## Describe your changes
We currently try to infer the transformers cache from environment
variables but those are not set most of the time. `transformers.utils`
has a variable which handles this logic internally so we can just use
that instead.

Also added some exception handling to remove the intermediate files that
the model builder adds to the cache dir. Ideally, the transformers cache
and intermediate cache should be dealt separately in order to not
pollute the transformers cache.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link