Restrict initialization of passes to what's called out in pass_flows
Config can list any number of passes, but the ones that are actually run
are the ones listed in pass_flows. This avoids the case where a pass may
not be able to successfully be initialized which shouldn't matter because
its not going to be run unless it is explicitly called out in the pass_flows.