ModelHandler: `cache_model` option in `load_model` (#1379)
## Describe your changes
Pytorch and Hf model handlers keep a reference to the load pytorch model
object for faster loading. This is helpful for use cases like evaluation
where we can use the loaded model for multiple metric runs.

However, some passes modify the model object in place, so these passes
have to set the `model.model` attribute to `None` so that the input
model doesn't use the modified model object. Adding a `cache_model`
option to the `load_model` method allows the model handler to decide
whether to cache the loaded model object or not. This makes the logic in
the passes safer.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link