GPTQ/AWQ: Support adapters and onnx export (#1390)
## Describe your changes
- GPTQ/AWQ have `pack_model_for_onnx_conversion` option which modifies
the quantization algorithm to pack the quantized model for onnx
conversion using `MatMulNBits` contrib op. However, the output model is
a pytorch model object which can only be used for onnx export. We cannot
apply any other passes on it. Also, the awq pass doesn't pack the
weights at all.
- Implemented a new functionality where the quantized models are
repacked right before export. This framework is extensible to newer
quantization methods which implement their own QuantLinear classes. The
output of the quantization passes are now hf models if the input is an
hf model.
- GPTQ/AWQ also support models with adapters by just quantizing the base
model and leaving the adapters as is in float precision.

**Note:** Support for QDQ export will be added in a follow-up PR.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link