ExtractAdapters: Default initializer for quantized modules (#1444)
## Describe your changes
Allow `optional_inputs` for quantized adapters also. The default
initializers have the same shape as the original weights but the scales
are made 0 so that the dequantized weight is 0. Quant weight and zero
point are also made 0 to be clean and consistent.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link