Model Splitting: Add cost model, Simplify split logic (#1456)
## Describe your changes
- In addition to `num_splits`, `CaptureSplitInfo` pass now supports cost
model. The cost model is a csv containing the cost per component of the
model. Currently, the only cost supported is the number of bytes.
- The split decision is made based on the `max_memory` config parameter.
This config parameter is temporary and will be moved into the
accelerator spec. The required changes are bigger than the scope of this
PR so it will be handled in a follow up PR.
- Added `generate-cost-model` CLI to generate cost models for
huggingface models. Also added some pre-generated cost models for phi
and llama models under `assets/cost_models`.
- `include_all_nodes` option has been removed from `SplitModel` pass. It
always includes all nodes. This also made the logic simpler.
- OnnxDAG:
- Updated logic to handle unused inputs/outputs for nodes that are left
as `""`. Found in contrib operators commonly.
  - Handle unnamed nodes. Node names are optional in the onnx spec.

## Checklist before requesting a review
- [x] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [x] Update documents if necessary.
- [x] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link
Fixes:
#1459