Phi3.5 example for NPU (#1680)
## Describe your changes
New example for preparing llm for NPU. The example uses
`phi-3.5-mini-instruct` but the model path can be replaced with other
models such as llama, phi-4, qwen2 and deepseek distilled models.

Changes made to enable this:
- New graph surgeries to remove rope multi cache and attention graph to
make the quantization simpler.
- New graph surgery to convert (Skip)SimplifiedLayerNormalization nodes
to L2Norm subgraph so it can run on NPU.
- Clean up context binary generator logic by removing the tempdir which
is not necessary.
- Replace `exclude_embeds` and `exclude_lm_head` options in
`CaptureSplitInfo` pass with `unique_embeds_lm_head_splits` which
assigns the embedding layer and lm head layers to their own splits.
- Fixed a bug in the accelerator creator where it allowed a workflow to
run even when the target system doesn't have the required EP installed.

The example list docs will be updated in a follow up PR since the doc
linkcheck requires the example folder to be present in the main branch
first.

## Checklist before requesting a review
- [x] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [x] Update documents if necessary.
- [x] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link
Closes #1630