Examples: Update genai API calls, Add deepseek/llama3 NPU examples (#1688)
## Describe your changes
- Updated the examples and docs with the latest version of ort genai api
calls
- Deepseek R1 distill and Llama 3.2 NPU examples added:
  - These just refer to the phi 3.5 example
- Added a note for the llama example that the quantized model quality
might not be good. Quality degradation appears to be worse than for the
phi or deepseek models. Need further investigation to see if it can be
improved.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link