Default accelerator is cpu is not specified (#1703)
## Describe your changes
- Update the behavior of the accelerator normalizer to choose cpu and
cpu-ep as the default accelerator if the target accelerator is not
specified.
- CPU EP is the default EP used when creating an ort session. It is also
the most generic EP.
- Otherwise, the workflow cycles through all EPs installed in the target
system. This behavior might have been desirable when multi-ep and
automatic accelerator+model selection was the focus but it's not
anymore. This behavior is confusing and requires the user to install the
correct ort package in the target system.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link