Create genai_config for NPU llms, inference example (#1699)
## Describe your changes
Fixes and updates:
- `CompositeModelHandler` now has an optional `model_path` resource to
store additional files. In practice, most of the composite models that
are output of passes use the same directory for the component models.
- Abstracted out the common llm-pipeline logic from the passes. It also
automatically updates the genai config with the correct model io and
session options.
  - Fixed untie word embedding implementation in hf model wrapper.
- Fixed the data type of the layernorm weight in the rms norm graph
surgeries. In numpy 2.0+, the data type after multiplication with
`sqrt_n` gets promoted to double. but we want float.
- Additional files carry over now supports composite models as well.
Removed unnecessary check for input model dir. It was a remnant from the
original implementation.

Examples:
- Instructions and example inference script added
- Qwen 2.5 example added.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link