HfModelHandler: fix condition for no dummy inputs (#1716)
## Describe your changes
Same fix as #1682. When the dummy inputs is a single tensor, `if
dummy_inputs` is ambiguous and leads to an error. Check for `not None`
instead.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link