Use ONNX IR in graph surgeries (#1751)
## Describe your changes

Use ONNX IR in graph surgeries to replace some proto manipulation. This
PR adds a `call_ir` in the base class that is called when `__call__` is
not overriden. This way the changes can be incremental and compatible
with existing usages of the `Surgeon` class.

Refactoring:
- Renamed the Surgeon class that operates on proto to ProtoSurgeon to
make the intention clear. Developers can subclass ProtoSurgeon to
operate on proto and subclass Surgion to transform graphs directly with
IR.

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [x] Make sure all tests can pass.
- [x] Update documents if necessary.
- [x] Lint and apply fixes to your code by running `lintrunner -a`
- [ ] Is this a user-facing change? If yes, give a description of this
change to be included in the release notes.
- [ ] Is this PR including examples changes? If yes, please remember to
update [example
documentation](https://github.com/microsoft/Olive/blob/main/docs/source/examples.md)
in a follow-up PR.

## (Optional) Issue link